moved detail row to own class
