+++ lib/widgets/overview.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:habit_tracker/blocs/habit_bloc/habit_bloc.dart';
 import 'package:habit_tracker/widgets/custom_card.dart';
+import 'package:habit_tracker/widgets/overview_detail_row.dart';
 import 'package:habit_tracker/widgets/spacer_box.dart';
 
 class Overview extends StatelessWidget {
@@ -37,11 +38,11 @@ class Overview extends StatelessWidget {
                     style: Theme.of(context).textTheme.headline5,
                   ),
                   SpacerBox.size16(),
-                  DetailRow(
+                  OverviewDetailRow(
                     detail: state.totalHabits.toString(),
                     text: 'total habits',
                   ),
-                  DetailRow(
+                  OverviewDetailRow(
                     detail: state.unfinishedHabits.toString(),
                     text: 'uncompleted habits',
                   )
@@ -70,35 +71,6 @@ class Overview extends StatelessWidget {
             )
           ],
         ))
-      ],
-    );
-  }
-}
-
-class DetailRow extends StatelessWidget {
-  final String detail;
-  final String text;
-  const DetailRow({
-    Key key,
-    this.detail,
-    this.text,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        SizedBox(
-          width: 40,
-          child: Text(
-            detail,
-            style: Theme.of(context).textTheme.headline6,
-          ),
-        ),
-        Text(
-          text,
-          style: Theme.of(context).textTheme.bodyText1,
-        )
       ],
     );
   }
